Split rr_nondet_log into replay and records
Previously, only one rr_nondet_log struct was in use, with a type
specifying whether its a record or a replay. This commit splits this
in two structs, rr_nondet_replay_log and rr_nondet_record_logs, to
enable records in replay.

Additionally, for records in records, the rr_nondet_record_logs are
stored in a linked list; rr_write_item takes now as additional argument
an RR_log* ptr. If this is NULL, the item is written to all record logs,
otherwise only to the one specified in ptr.

(Most likely), this commit does not break previous functionality, but
the desired functionality is still not achieved.